+++ src/core/Portfolio/Output/TransactionList.cs
@@ -60,7 +60,7 @@ namespace core.Portfolio.Output
         public IEnumerable<string> Tickers { get; set; }
         public IEnumerable<TransactionGroup> Grouped { get; set; } 
         
-        public double Credit => Transactions.Sum(t => t.Credit);
-        public double Debit => Transactions.Sum(t => t.Debit);
+        public decimal Credit => Transactions.Sum(t => t.Credit);
+        public decimal Debit => Transactions.Sum(t => t.Debit);
     }
 }
