+++ src/core/Stocks/View/StockDetailView.cs
@@ -5,7 +5,7 @@ namespace core.Stocks.View
     public class StockDetailsView
     {
         public string Ticker { get; internal set; }
-        public double Price { get; internal set; }
+        public decimal Price { get; internal set; }
         public CompanyProfile Profile { get; internal set; }
         public StockAdvancedStats Stats { get; internal set; }
     }
